Move terminal related functions to `term`
